修改未登录提示

+++ back/src/main/java/com/qianlei/zhifou/controller/AnswerController.java
@@ -1,5 +1,6 @@
 package com.qianlei.zhifou.controller;
 
+import com.qianlei.zhifou.common.AuthorizationException;
 import com.qianlei.zhifou.entity.Answer;
 import com.qianlei.zhifou.service.IAnswerService;
 import com.qianlei.zhifou.vo.AnswerVo;
@@ -18,17 +19,29 @@ public class AnswerController {
   }
 
   @PostMapping("/")
-  public Answer createAnswer(@RequestBody Answer answer, @CookieValue String token) {
+  public Answer createAnswer(
+      @RequestBody Answer answer, @CookieValue(required = false) String token) {
+    if (token == null) {
+      throw new AuthorizationException("用户未登录");
+    }
     return answerService.createAnswer(answer, token);
   }
 
   @PostMapping("/id/{answerId}/agree/")
-  public void agree(@PathVariable("answerId") Integer answerId, @CookieValue String token) {
+  public void agree(
+      @PathVariable("answerId") Integer answerId, @CookieValue(required = false) String token) {
+    if (token == null) {
+      throw new AuthorizationException("用户未登录");
+    }
     answerService.agree(answerId, token);
   }
 
   @DeleteMapping("/id/{answerId}/agree/")
-  public void removeAgree(@PathVariable("answerId") Integer answerId, @CookieValue String token) {
+  public void removeAgree(
+      @PathVariable("answerId") Integer answerId, @CookieValue(required = false) String token) {
+    if (token == null) {
+      throw new AuthorizationException("用户未登录");
+    }
     answerService.deleteAgree(answerId, token);
   }
 }
